Add files via upload
Since ancient times, since the ice Age, when the thickness of the ice sheet reached 2000 m to the present day, we can admire such an interesting state of matter as ice.   Scientists have calculated that the total mass of snow and ice on the whole Earth is approximately 2.423x10 (22 degrees) g.